third_party/WebKit/LayoutTests/bluetooth/service/getCharacteristic/gen-characteristic-not-found.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('HeartRateAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}]}))
    .then(device => device.gatt.connect())
    .then(gatt => gatt.getPrimaryService('heart_rate'))
    .then(service => assert_promise_rejects_with_message(
      service.getCharacteristic('battery_level'),
      new DOMException(
        'No Characteristics matching UUID ' + battery_level.uuid + ' found ' +
        'in Service with UUID ' + heart_rate.uuid + '.',
        'NotFoundError')));
}, 'Request for absent characteristics with UUID. Reject with NotFoundError.');

</script>
